requirements.txt
absl-py==2.4.0
asttokens==3.0.1
attrs==26.1.0
av==17.0.0
cffi==2.0.0
comm==0.2.3
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jax==0.9.2
jaxlib==0.9.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.13
ml_dtypes==0.5.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pose_format==0.12.3
prompt_toolkit==3.0.52
protobuf==4.25.9
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.9.0
PyYAML==6.0.3
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
simple-video-utils==0.0.6
six==1.17.0
sounddevice==0.5.5
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing_extensions==4.15.0
wcwidth==0.6.0

torchvision